Handle multiple failed scrapes gracefully.
Before if we couldn't get the pool name from the php status page, we set
the name to ""; which caused conflicts if more than one poor failed to
scrape.  We now set a default name based on the name of the socket, which
will be overridden by the configured pool name upon successful scrape.